terraform/modules/cloud_function/requirements.txt
functions-framework==3.5.0
psycopg2-binary==2.9.9